Move rename plate classes to separate module
