+++ .idea/runConfigurations/covid_rename_plates_rna.xml
@@ -12,7 +12,7 @@
     <option name="ADD_CONTENT_ROOTS" value="true" />
     <option name="ADD_SOURCE_ROOTS" value="true" />
     <option name="SCRIPT_NAME" value="$PROJECT_DIR$/clarity-ext-scripts/src/clarity-ext/clarity_ext/cli.py" />
-    <option name="PARAMETERS" value="--level INFO extension --cache False clarity_ext_scripts.covid.rename_plates_rna test" />
+    <option name="PARAMETERS" value="--level INFO extension --cache False clarity_ext_scripts.covid.rename_plates.rename_plates_rna test" />
     <option name="SHOW_COMMAND_LINE" value="false" />
     <option name="EMULATE_TERMINAL" value="false" />
     <option name="MODULE_MODE" value="false" />
