+++ clarity-ext-scripts/clarity_ext_scripts/covid/rename_plates/rename_plates_rna.py
@@ -1,5 +1,5 @@
-from clarity_ext_scripts.covid.rename_plates_base import Extension as RenamePlatesBase
-from clarity_ext_scripts.covid.plate_name_generator import InheritancePlateNameGenerator
+from clarity_ext_scripts.covid.rename_plates.rename_plates_base import Extension as RenamePlatesBase
+from clarity_ext_scripts.covid.rename_plates.plate_name_generator import InheritancePlateNameGenerator
 
 
 class Extension(RenamePlatesBase):
